+++ script/mousemove-v0.1.2.js
@@ -2,7 +2,7 @@
 // @brief: 			Automated mouse cursor for web presentation 
 // @author: 		Justin D. Byrne 
 // @email: 			justin@byrne-systems.com 
-// @version: 		0.1.0 
+// @version: 		0.1.2 
 // @license: 		GPL-2.0
  
 /**
@@ -506,6 +506,8 @@ class Cursor
 
 	//// 	[ POSITION ]    ////////////////////////////////
 
+		_getElementByXPath = ( path ) => document.evaluate ( path, document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null ).singleNodeValue;
+
 		/**
 		 * Set position property
 		 * @param 			{Point} point 								X & Y coordinates
@@ -542,23 +544,16 @@ class Cursor
 		 */
 		set angle ( id )
 		{
-			if ( document.getElementById ( id ) != null )
+			let _element = this.#_getElement ( id );
+
+
+			if ( _element != undefined )
 			{
-				let _element = document.getElementById ( id );
-
-				let _destination =
-				{
-					x: this.#_toNumber ( _element.style.left ) + ( this.#_toNumber ( _element.style.width )  / 2 ),
-
-					y: this.#_toNumber ( _element.style.top  ) + ( this.#_toNumber ( _element.style.height ) / 2 )
-				}
+				let _destination = this.#_getCenterPoint ( _element );
 
 
 		  		this.#_config.calculations.angle = Math.atan2 ( _destination.y - this.position.y, _destination.x - this.position.x );
 			}
-			else
-
-				console.log ( ` >> [ ERROR ]: Id ${id} is not a valid identifier !` );
 		}
 
 		/**
@@ -578,23 +573,16 @@ class Cursor
 		 */
 		set distance ( id )
 		{
-			if ( document.getElementById ( id ) != null )
+			let _element = this.#_getElement ( id );
+
+
+			if ( _element != undefined )
 			{
-				let _element = document.getElementById ( id );
-
-				let _point =
-				{
-					x: this.#_toNumber ( _element.style.left ) + ( this.#_toNumber ( _element.style.width  ) / 2 ),
-
-					y: this.#_toNumber ( _element.style.top  ) + ( this.#_toNumber ( _element.style.height ) / 2 )
-				}
+				let _point = this.#_getCenterPoint ( _element );
 
 
 				this.#_config.calculations.distance = Math.sqrt ( ( Math.pow ( _point.x - this.position.x , 2 ) ) + ( Math.pow ( _point.y - this.position.y, 2 ) ) );
 			}
-			else
-
-				console.log ( ` >> [ ERROR ]: Id ${id} is not a valid identifier !` );
 		}
 
 		/**
@@ -609,11 +597,35 @@ class Cursor
 	////    ( PRIVATE )    /////////////////////////////////
 
 		/**
+		 * Returns a DOM's element based on its identifier
+		 * @param  			{string} id 								CSS Identifier or xpath
+		 * @return 			{object}    								HTML DOM element
+		 */
+		#_getElement = ( id ) => ( document.getElementById ( id ) != null )
+
+								     ? document.getElementById ( id )
+
+								     : ( this._getElementByXPath ( id ) != null )
+
+									       ? this._getElementByXPath ( id )
+
+									       : undefined;
+
+		/**
+		 * Gets the center point of an element
+		 * @param  			{object} element 							HTML DOM element
+		 * @return 			{Point}         							X & Y Coordinates
+		 */
+		#_getCenterPoint = ( element ) => ( { x: element.offsetLeft + ( element.clientWidth  / 2 ),
+
+												   y: element.offsetTop  + ( element.clientHeight / 2 ) } );
+
+		/**
 		 * Converts CSS string value to number/integer
 		 * @param 			{string} value 								CSS string value in pixels
 		 * @return 			{number} 									Number value of parsed value
 		 */
-		#_toNumber ( value )
+		#_pxToNumber ( value )
 		{
 			return Number ( value.replace ( /px$/, '' ) );
 		}
@@ -645,9 +657,9 @@ class Cursor
 
 			this.position =
 			{
-				x: this.#_toNumber ( _image.style.left ),
+				x: this.#_pxToNumber ( _image.style.left ),
 
-				y: this.#_toNumber ( _image.style.top  )
+				y: this.#_pxToNumber ( _image.style.top  )
 			}
 		}
 
@@ -739,19 +751,12 @@ class Cursor
 		 */
 		toNextElement ( id )
 		{
-			let _element = document.getElementById ( id );
+			let _element = this.#_getElement ( id );
 
 
-			if ( document.getElementById ( id ) != null )
+			if ( _element != null )
 			{
-				let _element = document.getElementById ( id );
-
-				this.position =
-				{
-					x: this.#_toNumber ( _element.style.left ) + ( this.#_toNumber ( _element.style.width  ) / 2 ),
-
-					y: this.#_toNumber ( _element.style.top  ) + ( this.#_toNumber ( _element.style.height ) / 2 )
-				}
+				this.position = this.#_getCenterPoint ( _element );
 
 
 				this.#_mouseActions ( _element );
@@ -828,13 +833,17 @@ class MouseMove
 		{
 			hotkeys: [ 'ctrl+g', 'command+g' ] 					// For mousetrap
 		},
+		mousetrap:
+		{
+			cdn: '//cdnjs.cloudflare.com/ajax/libs/mousetrap/1.6.0/mousetrap.min.js'
+		},
 		about:
 	    {
 	        Author:    'Justin Don Byrne',
-	        Created:   'July, 27 2023',
+	        Created:   'Aug, 04 2023',
 	        Library:   'Mouse Move: Automated mouse cursor for web presentation',
-	        Updated:   'Aug, 04 2023',
-	        Version:   '0.1.0',
+	        Updated:   'Aug, 07 2023',
+	        Version:   '0.1.2',
 	        Copyright: 'Copyright (c) 2023 Justin Don Byrne'
 	    }
 	}
@@ -867,7 +876,7 @@ class MouseMove
 		 */
 		set cursor ( cursor )
 		{
-			this._cursor = ( cursor instanceof Cursor  ) ? cursor : this._cursor;
+			this._cursor = ( cursor instanceof Cursor  ) ? cursor     : this._cursor;
 
 			this._cursor = ( this._cursor == undefined ) ? new Cursor : this._cursor;
 		}